Don't try to remove the HTTP client from the loop after failure
It would croak with 'Notifier does not exist in collection', which still
correctly yields a failing future but isn't right.

I'm rather confused about how Net::Async::HTTP plays with the loop when
used with futures. I'm not convinced we're not leaking something
somewhere (see the fact that t/10-tester-http.t does not run under
no_pending_futures). To get to the bottom of that though I will need to
write some more controlled tests, read a lot of code and probably hassle
the author.