Take template config over member config
Rationale: templates are explicit in the config file, but members could
come from anywhere (eg Consul), and the group module will make some best
effort to fill those fields in.

If, as an admin, I've explicitly stated something in the config file,
then I don't want that to be overidden.

I can make the argument the other way, but we'll see if that comes up
before trying to implement some more complex conflict resolution system.